Define field mapping in funding case JSON schema
Mapping of submitted recipient contact ID of funding case to the
`recipient_contact_id` field of the `FundingCase` entity is now
specified in JSON schema.
