+++ tests/phpunit/Civi/Funding/FundingCase/Handler/FundingCaseFormUpdateValidateHandlerTest.php
@@ -25,7 +25,7 @@ use Civi\Funding\EntityFactory\FundingProgramFactory;
 use Civi\Funding\Form\FundingCase\FundingCaseValidationResult;
 use Civi\Funding\Form\FundingCase\FundingCaseValidatorInterface;
 use Civi\Funding\FundingCase\Command\FundingCaseFormUpdateValidateCommand;
-use Civi\Funding\Mock\FundingCaseType\FundingCase\Validation\TestFundingCaseValidatedData;
+use Civi\Funding\Mock\FundingCaseType\FundingCase\Validation\ValidatedFundingCaseDataMock;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
@@ -56,7 +56,7 @@ final class FundingCaseFormUpdateValidateHandlerTest extends TestCase {
     $fundingCase = FundingCaseFactory::createFundingCase();
 
     $data = ['foo' => 'bar'];
-    $validatedData = new TestFundingCaseValidatedData([]);
+    $validatedData = new ValidatedFundingCaseDataMock([]);
     $errorMessages = ['/a/b' => ['error']];
     $validationResult = FundingCaseValidationResult::newInvalid($errorMessages, $validatedData);
