+++ tests/phpunit/Civi/Funding/Mock/FundingCaseType/FundingCase/Validation/TestFundingCaseValidator.php
@@ -24,8 +24,10 @@ use Civi\Funding\Entity\FundingCaseTypeEntity;
 use Civi\Funding\Entity\FundingProgramEntity;
 use Civi\Funding\Form\FundingCase\AbstractFundingCaseValidator;
 use Civi\Funding\Form\FundingCase\FundingCaseValidationResult;
+use Civi\Funding\Form\FundingCase\ValidatedFundingCaseData;
 use Civi\Funding\Mock\FundingCaseType\Traits\TestSupportedFundingCaseTypesTrait;
 use Civi\RemoteTools\JsonSchema\JsonSchema;
+use Civi\RemoteTools\JsonSchema\Validation\ValidationResult;
 
 final class TestFundingCaseValidator extends AbstractFundingCaseValidator {
 
@@ -40,10 +42,13 @@ final class TestFundingCaseValidator extends AbstractFundingCaseValidator {
     FundingCaseEntity $fundingCase,
     array $formData,
     JsonSchema $jsonSchema,
-    array $validatedData,
+    ValidationResult $jsonSchemaValidationResult,
     int $maxErrors
   ): FundingCaseValidationResult {
-    return FundingCaseValidationResult::newValid(new TestFundingCaseValidatedData($validatedData));
+    return FundingCaseValidationResult::newValid(new ValidatedFundingCaseData(
+      $jsonSchemaValidationResult->getData(),
+      $jsonSchemaValidationResult->getTaggedData()
+    ));
   }
 
   /**
@@ -54,10 +59,13 @@ final class TestFundingCaseValidator extends AbstractFundingCaseValidator {
     FundingCaseTypeEntity $fundingCaseType,
     array $formData,
     JsonSchema $jsonSchema,
-    array $validatedData,
+    ValidationResult $jsonSchemaValidationResult,
     int $maxErrors
   ): FundingCaseValidationResult {
-    return FundingCaseValidationResult::newValid(new TestFundingCaseValidatedData($validatedData));
+    return FundingCaseValidationResult::newValid(new ValidatedFundingCaseData(
+      $jsonSchemaValidationResult->getData(),
+      $jsonSchemaValidationResult->getTaggedData()
+    ));
   }
 
 }
